rMalMon1 EAR
Made several changes to the assembly, remapped the HiC and further re-ordered some of the unlocs of SUPER_W on the latest pretext. The latest pretext with Z and W next to one another is savestate_2. savestate_0 corresponds to pretext file before the modifications. Hope this is fine now and we apologize for the delay.