Amend Final qqEusFeti1 EAR
we noticed that an unplaced sequenced miss-labelled as "SUPER_34". It was so tiny that we don't even see it in our pre and post-curation pretexts... it has a length 137,230 that is very close to the texel resolution  133,416.392761 bp/texel

The EAR submitted during review states an n=34 while in fact we just painted 33 super-scaffolds (n=33). For consistency, I think the corrected EAR attached should replace the current one